refactor(modal): 优化modal目录结构
